Rewrite arch() fallback code in package-manager lib.
* libs/package-manager.bash.in (arch): Complete rewrite of the
fallback code if there is no make.profile symlink. It now relies
on bash variables HOSTTYPE and OSTYPE which are derived from the
standard GNU triplet. The previous implementation had used uname
which would query the kernel, while we need the userland. This
fixes a problem on ppc with 64 bit kernel and 32 bit userland.
Thanks to Brent Baude <ranger@gentoo.org> and Raúl Porcel
<armin76@gentoo.org> for the suggestion.